requirements.txt
groq
openai          
requests      
python-docx   
SpeechRecognition
pyaudio
pyttsx3    
python-dotenv
openpyxl